# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple password generator API built with Go and Gin.",
        "title": "EBVN Bookmark API",
        "contact": {},
        "version": "1.4"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/bookmarks/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing bookmark's description and URL. Only the bookmark owner can update it.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookmark"
                ],
                "summary": "Update a bookmark",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bookmark ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated bookmark details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_bookmark.updateBookmarkInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_internal_dto.SuccessResponse-any"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_pkg_response.Message"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_pkg_response.Message"
                        }
                    },
                    "404": {
                        "description": "Bookmark not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_pkg_response.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_pkg_response.Message"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an existing bookmark. Only the bookmark owner can delete it.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookmark"
                ],
                "summary": "Delete a bookmark",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bookmark ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_internal_dto.SuccessResponse-any"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_pkg_response.Message"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_pkg_response.Message"
                        }
                    },
                    "404": {
                        "description": "Bookmark not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_pkg_response.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_pkg_response.Message"
                        }
                    }
                }
            }
        },
...
        "/v1/links/redirect/{code}": {
            "get": {
                "description": "Retrieve the original URL for a short code and redirect the client",
                "tags": [
                    "URL"
                ],
                "summary": "Redirect to original URL",
                "parameters": [
                    {
                        "type": "string",
                        "example": "abc1234",
                        "description": "Short code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirects to the original URL"
                    },
                    "400": {
                        "description": "Bad Request - wrong format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_pkg_response.Message"
                        }
                    }
                }
            }
        },
...
        "/v1/users/login": {
            "post": {
                "description": "Authenticate a user with username and password, returns a JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Login a user",
                "parameters": [
                    {
                        "description": "User login credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_user.loginInputBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_internal_dto.SuccessResponse-string"
                        }
                    },
                    "400": {
                        "description": "Invalid username or password",
                        "schema": {
                            "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_pkg_response.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_pkg_response.Message"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_HadesHo3820_ebvn-golang-course_internal_dto.Metadata": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer",
                    "example": 1
                },
                "first_page": {
                    "type": "integer",
                    "example": 1
                },
                "last_page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 10
                },
                "total_records": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "github_com_HadesHo3820_ebvn-golang-course_internal_dto.SuccessResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_internal_dto.Metadata"
                }
            }
        },
...
        "github_com_HadesHo3820_ebvn-golang-course_internal_dto.SuccessResponse-string": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_internal_dto.Metadata"
                }
            }
        },
...
        "github_com_HadesHo3820_ebvn-golang-course_pkg_response.Message": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "Details contains a list of specific error messages or additional information, if any.\nIf empty, this field is omitted from the JSON response."
                },
                "message": {
                    "description": "Message is a brief summary of the response (e.g., \"Input error\").",
                    "type": "string"
                }
            }
        },
...
        "internal_handler_bookmark.updateBookmarkInput": {
            "type": "object",
            "required": [
                "id",
                "url"
            ],
            "properties": {
                "description": {
                    "description": "Description of the bookmark",
                    "type": "string",
                    "maxLength": 255,
                    "example": "Google"
                },
                "id": {
                    "description": "ID is the bookmark identifier from the URL path",
                    "type": "string"
                },
                "url": {
                    "description": "URL to be shortened",
                    "type": "string",
                    "maxLength": 2048,
                    "example": "https://www.google.com"
                }
            }
        },
...
        "internal_handler_user.loginInputBody": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "Password is the user's password.",
                    "type": "string",
                    "minLength": 8
                },
                "username": {
                    "description": "Username is the user's login identifier.",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your Bearer token in the format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}