# internal/handler/url/get_url.go
package url

import (
	"errors"
	"net/http"
	"strings"

	"github.com/HadesHo3820/ebvn-golang-course/internal/service"
	"github.com/HadesHo3820/ebvn-golang-course/pkg/response"
	"github.com/gin-gonic/gin"
	"github.com/rs/zerolog/log"
)

// GetUrl handles HTTP GET requests to retrieve and redirect to the original URL.
// It extracts the short code from the URL path, validates it, queries the service
// layer for the original URL, and redirects the client using HTTP 302 Found.
//
// Path Parameters:
//   - code: The 7-character alphanumeric short code generated by ShortenUrl.
//
// Responses:
//   - 302 Found: Redirects to the original URL.
//   - 400 Bad Request: Code is empty, malformed, or does not exist.
//   - 500 Internal Server Error: Database or service layer failure.
//
// @Summary Redirect to original URL
// @Description Retrieve the original URL for a short code and redirect the client
// @Tags URL
// @Param code path string true "Short code" example(abc1234)
// @Success 302 "Redirects to the original URL"
// @Failure 400 {object} map[string]string "Bad Request - wrong format"
// @Failure 500 {object} response.Message
// @Router /v1/links/redirect/{code} [get]
func (h *urlHandler) GetUrl(c *gin.Context) {
	// Extract the short code from the URL path parameter.
	// The route is defined as /v1/links/redirect/:code, so Gin parses the dynamic segment.
	// Note: Gin may include leading/trailing slashes (e.g., "/abc1234/") for requests like /v1/links/redirect/abc1234/
	code := c.Param("code")

	// Trim any leading/trailing slashes that Gin might include from the URL path.
	code = strings.Trim(code, "/")

	if code == "" {
		c.JSON(http.StatusBadRequest, gin.H{"message": "wrong format"})
		return
	}
	// Query the service layer to retrieve the original URL for this code.
	url, err := h.urlService.GetUrl(c, code)
	if err != nil {
		// Check if the error is specifically "code not found" using sentinel error comparison.
		// This allows us to return a 400 Bad Request instead of a generic 500 error.
		if errors.Is(err, service.ErrCodeNotFound) {
			c.JSON(http.StatusBadRequest, gin.H{"message": "url not found"})
			return
		}

		// For any other errors (e.g., database connection issues), return 500.
		log.Error().
			Str("code", code).
			Err(err).
			Msg("Failed to retrieve URL for short code")
		c.JSON(http.StatusInternalServerError, response.InternalErrResponse)
		return
	}

	// Redirect the client to the original URL using HTTP 302 (Found).
	// This is the standard behavior for URL shorteners.
	c.Redirect(http.StatusFound, url)
}